popup switched to its own class, button fragment calls now call the right fragments and instructions started
