+++ app/src/main/java/com/example/fridgeapprator/database/IHSDatabase.java
@@ -20,7 +20,7 @@ import com.example.fridgeapprator.model.ShoppingList;
 import com.example.fridgeapprator.model.ShoppingListProduct;
 
 @TypeConverters({Converters.class})
-@Database(entities = {Product.class, ProductType.class, ShoppingList.class, ShoppingListProduct.class}, version = 44, exportSchema = false)
+@Database(entities = {Product.class, ProductType.class, ShoppingList.class, ShoppingListProduct.class}, version = 45, exportSchema = false)
 public abstract class IHSDatabase extends RoomDatabase {
 
     public abstract ProductDao productDao();
