+++ app/src/main/java/com/example/fridgeapprator/fragment/InstructionsFragment.java
@@ -1,0 +1,22 @@
+package com.example.fridgeapprator.fragment;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
+
+import com.example.fridgeapprator.R;
+
+public class InstructionsFragment extends Fragment {
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+
+        View view = inflater.inflate(R.layout.instructions_fragment, container, false);
+
+        return view;
+
+    }
+}
