+++ app/src/main/res/layout/shopping_list_fragment.xml
@@ -64,11 +64,12 @@
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recyclerViewShoppingList"
-        android:layout_width="412dp"
+        android:layout_width="0dp"
         android:layout_height="400dp"
         android:layout_marginTop="47dp"
         android:layout_marginBottom="10dp"
         app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/inputNewProductTypeName"
         tools:listitem="@layout/fridge_item_list_fragment" />
